Procesado_de_feature: se agrego el script que no escala, se cambiaron de directorio, se corrigieron las rutas de la BD y guardado
